Add keyboard focus hints to KCardPile.
These allow a pile to give hints about which cards can, cannot and
should be focused. The implementation works quite well, but I'm not in
love with the public API.

My goal here is not to require the individual games to set up a bunch of
stuff in order to support keyboard control, but instead to allow the
games to augment the user experience by providing additional information
to the keyboard control mechanisms.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1125477 283d02a7-25f6-0310-bc7c-ecb5cbfe19da