Improve handling of draw, deal row and redeal events.
Previously, each DealerScene subclasses newCards() method had ugly code
to return a non null pointer if the method was called while an animation
was running. This allowed us to ignore these requests without worrying
about crashing the demo. With the new animationDone() framework, we can
now guarantee the these methods aren't called while animations are
running.

As a bonus, requests to draw, deal row or redeal are now delayed until
the end of the animation instead of just silently ignored.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1134180 283d02a7-25f6-0310-bc7c-ecb5cbfe19da