Add hidden actions for starting the next or previous numbered deal.
An advanced user requested a button to automatically open the next
numbered deal. I was about to tell him that I'd rather not clutter up
the UI with a feature that will be used by a very, very small percentage
of users, but I then realised that thanks to KDE's fancy action system,
I didn't necessarily have to clutter up the UI to add this feature.

So this commit adds actions "Next Deal" and "Previous Deal", but the
actions appear neither in the menus, nor in the default toolbar. Users
wanting access to this feature can either use them via their shortcuts
(Ctrl+Plus and Ctrl+Minus) or by adding them to their toolbar.

BUG:244543

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1150426 283d02a7-25f6-0310-bc7c-ecb5cbfe19da