Add a KGameRenderedPixmapItem class as a simpler alternative to KGRI.
While it has some fancy features, KGameRenderedItem's complexity can
also prove a hindrance. It doesn't provide access to the full QGPI
API. It adds the wait of QObject to every item (which is probably not a
big deal, but it bugs me.) One cannot override the implementation of
paint() because of the secret child item. Etc.

For the simple needs of Killbots all that is really needed is a
QGPI that can automatically fetch its pixmap from the KGR. This simple,
simple, simple class provides that.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1156146 283d02a7-25f6-0310-bc7c-ecb5cbfe19da