Remove caching of the theme aspect ratio and cursors from Renderer.
Caching the theme aspect ratio is no longer really neccessary now that
we have caching in KGameRenderer::boundsOnSprite().

When I first added the QCursor caching I did so just because I was
unsure of whether constructing cursors from existing pixmaps was
expensive or not. I just did some profiling and it seems that it is not
(on X11, at least).

Text color remains cached as it involves pixmap to image conversion and
is called by NumericDisplayItem::paint() which can occurs tens of times
per second.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1164843 283d02a7-25f6-0310-bc7c-ecb5cbfe19da