Replace KCardPile::availableSpace with KCardScene::spaceAllottedToPile.
Storing the availableSpace in the KCardPile itself was ugly as it
required either KCardPile::setAvaibleSpace to be public or KCardScene
to be a friend. Logically, I think it makes more sense for the scene
to keep track of that information anyway, as the pile itself has no
control over it.

Also, the QHash was already being filled as part of the calculations
inside KCardScene::layoutPiles so now we just hold onto it and use it
as a cache.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1168070 283d02a7-25f6-0310-bc7c-ecb5cbfe19da