Remove KGameRendererClient::pixmap().
As far as I can tell this method (and corresponding data member) are
just holdovers from a previous version of the API. (Or possibly it's a
not yet implemented feature?) Fortunately, no one was using it, so we
can safely remove it.

I only noticed its existance because it makes call to KGameRenderedItem
::pixmap() ambiguous as both parent classes contain a pixmap() method.
So if this method is ever readded to the API, a name change may be
advisable.

CCMAIL:majewsky@gmx.net

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1169222 283d02a7-25f6-0310-bc7c-ecb5cbfe19da