Use a custom function for generating ID list in DealerScene subclasses.
For historical reasons KPat generates its card lists in a slightly
unusual order. Up until this point I was using that same order in
LibKCardGame. This seems like a bad idea because: a) It's probably
not what people expect. b) It's a goal to get as many KPat specific
quirks out of LibKCardGame.

So now LibKCardGame use the more logical ordering and KPat does its
own thing.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1171451 283d02a7-25f6-0310-bc7c-ecb5cbfe19da