Create a common.cpp file to hold cache related functions.
Other common functions and/or constants will be moved there in the future.

 * createCache() is a function for creating KImageCache objects, to
guarantee that all uses of KImageCache in LibKCardGame are consistent.
This will be useful mostly when porting KCardThemeWidget to KImageCache.

 * cacheFind() is a template function that allows one to extract any
type from a KSharedDataCache provided it has an operator<< method for
QDataStream.

 * cacheInsert() is a template function that allows one to insert any
type into a KSharedDataCache provided it has a operator>> method for
QDataStream.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1171458 283d02a7-25f6-0310-bc7c-ecb5cbfe19da