Use integers for sceneRect calculations.
Previously we would often see a white strip a single pixel wide on
the right and/or bottom edge of the view. This was caused by QRectFs
being rounded to QRects for the actual painting to screen pixels. By
ensuring that the screneRect is composed of whole numbers, we eliminate
this issue.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1175046 283d02a7-25f6-0310-bc7c-ecb5cbfe19da