fix logical error
Checking this option ensures that the created games are solvable.
The option label was wrong. I could have kept the lebel and
adjusted the code accordingly but I decided to keept the code and
switch the meaning of the label, since it makes more sense
this way.

Thanks to all the reporters for analysing this. :)

BUG: 145953


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1181702 283d02a7-25f6-0310-bc7c-ecb5cbfe19da