Adapt Kolf::Overlay and Kolf::Shape from playground/games/kolf2/base.
I'm aiming to restrict CanvasItem subclasses to those which are actually
relevant for the gameplay (e.g. Bumper, Cup, Bridge; e.g. NOT RectPoint
and WallPoint). The first step is to introduce a means of displaying
editing interfaces without the need for additional CanvasItems, i.e.
Kolf::Overlay.

The Overlay class relies on the Shape class which determines the
outlines, but Shape will also be useful for the upcoming step to Box2D.
That code is already available in playground/games/kolf2/base, but I've
kept things simple in this first step. The aim is to replace the editing
interface first, and further simplify datastructures in the various
CanvasItem subclasses in the process.

Currently, an overlay is only installed on Cup and Wall. I plan to
stabilize those first before I install overlays on all items.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1191708 283d02a7-25f6-0310-bc7c-ecb5cbfe19da