Make Klickety do a more convincing job when pretending to be KSame.
Which this patch, "klickety --KSameMode" becomes very hard to
distinguish from "ksame". Window titles, menu entries, icons, the
about dialog, and highscore window all refer to Same Game.

I copied authors and credits from KSame to be shown when in KSame mode,
as I thought the was reasonable. The copyright of course is untouched as
the actual code is 100% Klickety's. If anyone disagrees with these
credits, please feel free to change them. This was just what I felt
would be best.

CCMAIL:shuizhuyuanluo@126.com

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1192352 283d02a7-25f6-0310-bc7c-ecb5cbfe19da