Rework the DealerInfo class to improve robustness.
This may be overkill, but it makes the DealerInfo class much easier to
use. Translation duties have been moved inside the class to ensure the
developer gets a translated string unless he explicitly requests the
untranslated bytes. The base ID and the subtype IDs are handled
separately inside the class and there are explicit getters for each.

BUG:252586

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1193161 283d02a7-25f6-0310-bc7c-ecb5cbfe19da