Make cards check that their size is correct before painting.
If a card is asked to paint() but it's current pixmap differs in size
from the current deck card size, it now explicitly requests a new a new
pixmap from the deck.

This will hopefully fix all the recent issues with cards having the
wrong size or not appearing at all. Technically, they were always
appearing, but sometimes they were appearing at a size of 0 by 0 pixels.
:)

BUG:255933

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1212966 283d02a7-25f6-0310-bc7c-ecb5cbfe19da