Move the pile layout animation from KCardPile to KCardScene.
This was the last bit of animating code living outside of KCardScene,
causing some ugly workarounds to get specialised behaviour. This commit
doesn't fixed all those, but it should make it easier to do so in the
future.

KCardPile is now only responsible for telling the KCardScene where it
thinks its cards should be placed (in layout space) and the scene
performs the animating and makes adjustments if things don't fit. All
in all this should be much cleaner.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1218289 283d02a7-25f6-0310-bc7c-ecb5cbfe19da