Convert KCardPile::spread() to a QPointF and remove cardOffset().
Whether to use a QSizeF or a QPointF for spread() is somewhat arbitrary
as neither one makes a whole lot of sense semantically, but a QPointF is
quite a bit easier to work with.

The cardOffset() method was mostly being used to be able to reduce the
spread on face down cards, but with the new cardPositions() virtual,
this somewhat KPat specific functionality can be easily moved into
PatPile.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1218291 283d02a7-25f6-0310-bc7c-ecb5cbfe19da