Remove usage of KIO.
KIO was used very inconsistently at only few places. It should be
possible to implement KIO on top of the new multithreaded Puzzle
class, but this requires one to wrap one's mind around it, and I do
not have the time to do that. So IMO it's better to remove the few
usages of KIO to send a clear message to the user about the feature
set that he can expect.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1225957 283d02a7-25f6-0310-bc7c-ecb5cbfe19da