Simplify selection of a pending local game
since we already enter the wanted ruleset when starting a local
game, only show pending games with that ruleset. When the user
decides to start a new game from the pending games list, also use
that ruleset. Until now the user had to select the ruleset twice
when starting a new local game with pending games present.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1230052 283d02a7-25f6-0310-bc7c-ecb5cbfe19da