check for data consistency when restarting a suspended game
When restarting a suspended game, check if the two data bases of
server and client agree about number of played hands and player balances.
If they do not match, show a warning and abort the game.
There is currently no way to fix this situation unless you want
to edit the data bases directly. If this ever happens in wild
life, we might need to synchronize up to five data bases
(4 clients and one server)

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1230542 283d02a7-25f6-0310-bc7c-ecb5cbfe19da