Revert r1231677
That commit resizes the tiles in the tileset preview (config/tileset
selector) whenever the user selects a different tileset.
Since that commit, it is no more possible to edit a ruleset name in
the ruleset editor. This is very strange, I can see no connection
between this commit and the effect.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1233543 283d02a7-25f6-0310-bc7c-ecb5cbfe19da