do not kill the local server when it is still used
the client starts a local server if needed. Until now, when it did,
it would also kill it when ending. But meanwhile a different client
might use the same server. New: The client never kills the server.
But when the last client disconnects from a local server, it ends
itself.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1234720 283d02a7-25f6-0310-bc7c-ecb5cbfe19da