improve framework for testing AI variants
this is not a bug fix, but no normal user will execute this code so
it should be safe to add. This is only of interest for somebody
wanting to improve the robot player artificial intelligence.

In demo mode, the csv records for game results now have a new first
field holding the AI variant used for the human player.
Adapted kajonggtest.py and made it more tolerant against missing
single games

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1235790 283d02a7-25f6-0310-bc7c-ecb5cbfe19da