Suppress graphical artifact
when a player claimed 'no choice', some rest of the
discarded tile was still visible. Clicking on it
or resizing the window removed it. Since I do not
know why, I just do a resizeEvent for the main window
in this situation - this gets rid of the artifact.

This only happened with animated tile movement
disabled.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1251567 283d02a7-25f6-0310-bc7c-ecb5cbfe19da