add checks which avoid processing messages meant for an old hand
there seems to be a bug where a message from a client is sent
to the server only after the current hand ended and a new hand
has started - leading to a situation where east has 15 tiles
at the beginning of a hand. This commit does not fix that, but
it hopefully throws an exception if a message is received that
is not meant for the current hand.

the debug output --showtraffic suppresses displaying this token
which is sent with every message.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1273154 283d02a7-25f6-0310-bc7c-ecb5cbfe19da