Don't bother saving state on close if the player has already won.
This was a behaviour regression caused by the clean up of the loading
and saving code. Somewhat surprisingly, saving after the game has been
won worked without issue and loading that save caused the player to
instantly win again, so it acted like the state was actually remembered.
Of course, the user doesn't expect to fire up KPat just to be instantly
told that they've won.

CCBUG:291959

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1275129 283d02a7-25f6-0310-bc7c-ecb5cbfe19da