Rewrite the state saving and restoring code to be diff based.
Previously, KPat stored the state of every single card for every state
in its undo/redo history. Since the vast majority of moves involve less
than 10 cards and a decent percentage involve only one, this approach
is obviously quite inefficient.

As of this commit, KPat now stores only the old and new states of each
card that has been altered since the last move. This is both more
processor and memory friendly, which is great, but it also opens the
door to creating a new saved game format that stores the entire game
history.

DIGEST:

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1280161 283d02a7-25f6-0310-bc7c-ecb5cbfe19da