Allow autodropping only when the redo stack is empty.
If there are states in the redo stack, there's a pretty good chance
that the user would like to keep them, but autodropping would replace
them with a trivial move. This isn't an issue in normal play, but can
come up when the user plays with autodrop disabled, misses some drops,
enables autodrop and then undoes back a few states.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1280178 283d02a7-25f6-0310-bc7c-ecb5cbfe19da