declaring original call automatically discards the selected tile
it makes program logic much simpler if this is atomic. Before, it
was possible to declare OC and then to discard a tile which does not
result in a calling hand. While this is a legal thing to do, it would
probably be a mistake (unless your main objective is to confuse the
other players). One of the next commits will disallow declaring OC
if no claiming hand results.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1280937 283d02a7-25f6-0310-bc7c-ecb5cbfe19da