Get rid of DealerScenePrivate.
I don't know why DealerScene had a d-pointer, but it bugged me that it
was the only class in KPat that did. Getting rid of it just simplifies
the code that much further.

I also renamed a few variables when I moved them into DealerScene, just
to make their purposes a bit more obvious or to make them consistent
with changes in implementation of terminology.

Lastly, member variable initialisations have been moved into an
initialisation list.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1281290 283d02a7-25f6-0310-bc7c-ecb5cbfe19da