Support puzzles that contain more than one large NxNxN cube.
Change the colors and sizes used for highlighted, selected and unhighlighted cells.

Rewrite the highlighting and selection mechanism so as to highlight only the large cube or cubes that contain the selected cell, shrinking and darkening the cubes that are outside that region.

Avoid flickering of the picture as the cursor moves around.

Use the SKGraph class to provide all puzzle-layout data. Move internal data items from public to private. Use m_ for data names. Tidy up some of the indentation and parenthesising.


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1286150 283d02a7-25f6-0310-bc7c-ecb5cbfe19da