demo mode: fix duplicate code execution
AI was asked twice which tile to discard. This only matters since recently
because since then the random generator is used when two tiles have the same
uselessness. But we want the random generator to be called exactly with the
same order/frequency for all modes (nogui, gui, autoplay) so we can
easier reproduce bugs.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1294797 283d02a7-25f6-0310-bc7c-ecb5cbfe19da