Improve scoring explanation
when I have to discard a tile, I have 14 tiles which was explained as
a long hand with score 0. Now Kajongg explains score for the hand
without the focussed tile. So when moving the cursor between tiles,
the scoring explanation and the current total shown on the wall
adjust immediately.
Also, suppress output for players with score 0

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1297823 283d02a7-25f6-0310-bc7c-ecb5cbfe19da