Sometimes clients did not write into data base when they should
This happened when two clients used the same path for their database,
even if it was on different computers.
Until now, identity of data bases was checked by path and IP address,
but that worked only in my local network.
Now each kajongg data base has a unique ID.
Background: It is legal having two kajongg instances running on one
client - and if that happens, they will use the same database. So only
one of them should save game info in that data base. The server says who

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1300847 283d02a7-25f6-0310-bc7c-ecb5cbfe19da